+++ projects/ui/src/lib/components/index.ts
@@ -27,6 +27,7 @@ export * from './loader';
 export * from './menu';
 export * from './number-input';
 export * from './otp-input';
+export * from './pagination';
 export * from './popover';
 export * from './progress';
 export * from './radio';