+++ projects/ui/src/lib/components/pagination/index.ts
@@ -1,0 +1,2 @@
+export * from './kui-pagination.component';
+export * from './kui-pagination-variant.type';